requirements.txt
flask
flask-sqlalchemy


langchain-core
langchain-community
langchain-text-splitters
langchain-openai
langchain-google-genai
langchain-chroma

openai
google-generativeai

chromadb

pydantic

python-dotenv
werkzeug

psycopg2

pypdf